+++ bundlers/browser/BrowserMultiPack.ts
@@ -241,13 +241,11 @@ module BrowserMultiPack {
             var dst = bundleStreams[idx].stream;
             var opts = bundles.bundles[idx];
             var prelude = preludes[idx];
-            if (lineNumAry[idx] == null) { lineNumAry[idx] = 1 + StringUtil.countNewlines(prelude); }
 
-            var wrappedSrc = createWrappedSourceAndMap(row, sourceMaps, idx, lineNumAry[idx], prelude, preludePaths[idx], opts, firsts[idx]);
+            var wrappedSrc = createWrappedSourceAndMap(row, sourceMaps, lineNumAry, idx, prelude, preludePaths[idx], opts, firsts[idx]);
 
             var fullSrc = wrappedSrc.join("");
             dst.push(Buffer.from(fullSrc));
-            lineNumAry[idx] += StringUtil.countNewlines(fullSrc);
 
             firsts[idx] = false;
             if (row.entry && row.order !== undefined) {
@@ -264,19 +262,21 @@ module BrowserMultiPack {
     /** Create the source text for a given 'ModuleDepRow' that can be concatenated together with
      * other calls to this function to form a the source text for a single bundle file.
      * @param row the row with 'source' and other properties
-     * @param sourceMaps an array of bundle source maps, the source map for this bundle is located at 'sourceMapIndex', however it may be null
-     * if it is null, a source map can be created and assigned to the 'sourceMapIndex' in the array.
-     * @param sourceMapIndex the 'sourceMaps' index at which the source map for this bundle is located/stored
-     * @param lineNum the current line number count within the bundle being built
+     * @param sourceMaps an array of bundle source maps, the source map for this bundle is located at 'bundleIndex', however it may be null
+     * if it is null, a source map can be created and assigned to the 'bundleIndex' in the array.
+     * @param lineNumbers array of line numbers counts, the current line number of the bundle being built is located at 'bundleIndex'
+     * @param bundleIndex the 'sourceMaps' and 'lineNumbers' index at which the source map and current line number for this bundle are located/stored
      * @param prelude the 'prelude' source text for this bundle
      * @param preludePath the 'prelude' file path (can be fake)
      * @param opts the options for this bundle
      * @param first whether this is the first 'row' being added to this bundle (if it is the 'prelude' should be included) in the returned text
      * @returns string array containing the source text to insert into the bundle for this 'row'
      */
-    export function createWrappedSourceAndMap(row: ModuleDepRow, sourceMaps: CombineSourceMap.Combiner[], sourceMapIndex: number, lineNum: number,
+    export function createWrappedSourceAndMap(row: ModuleDepRow, sourceMaps: CombineSourceMap.Combiner[], lineNumbers: number[], bundleIndex: number,
         prelude: string, preludePath: string, opts: BundleDst & BrowserPackOptions, first: boolean
     ): string[] {
+        const initialLineNum = lineNumbers[bundleIndex] || 0;
+        var lineNum = initialLineNum;
         var wrappedSrc: string[] = [];
 
         // create the prelude text at the beginning of the bundle
@@ -290,6 +290,8 @@ module BrowserMultiPack {
                 wrappedSrc.push(pre, "=");
             }
             wrappedSrc.push(prelude, "({");
+
+            lineNum += StringUtil.countNewlines(prelude) + 1; // +1 for next line at which the next file will start appending
         }
         // or just a comma ',' to separate files within a bundle
         else {
@@ -298,13 +300,13 @@ module BrowserMultiPack {
 
         // update the source mapping
         if (row.sourceFile && !row.nomap) {
-            var sourceMap = sourceMaps[sourceMapIndex];
+            var sourceMap = sourceMaps[bundleIndex];
             if (!sourceMap) {
                 sourceMap = CombineSourceMap.create(null, opts.sourceRoot);
-                sourceMaps[sourceMapIndex] = sourceMap;
+                sourceMaps[bundleIndex] = sourceMap;
                 sourceMap.addFile(
                     { sourceFile: preludePath, source: prelude },
-                    { line: 0 }
+                    { line: initialLineNum }
                 );
             }
             sourceMap.addFile(
@@ -312,6 +314,9 @@ module BrowserMultiPack {
                 { line: lineNum }
             );
         }
+
+        lineNum += StringUtil.countNewlines(row.source) + 2; // +2 for the UMD function and dependency definition lines surrounding the source, see below
+        lineNumbers[bundleIndex] = lineNum;
 
         // the actual code wrapped in a prelude require function
         wrappedSrc.push(