+++ test/test-proj/App.ts
@@ -7,12 +7,19 @@ function main() {
         name: "test-proj",
         version: "0.1.0",
         description: "parser/bundler test project",
+        dependencies: { "base": "none" },
     };
+    HelperUtil.Ui = WidgetUi;
 
-    console.log(WidgetUi.createWidget());
+    var widget = WidgetUi.createWidget(HelperUtil.App);
+    console.log(widget);
+
     for (var collName in DataSource.collectionCache) {
         console.log(collName + ": " + DataSource.collectionCache[collName]);
     }
+
+    WidgetUi.updateWidget(widget);
+    console.log("done");
 }
 
 main();