+++ test/test-proj/helpers/HelperUtil.js
@@ -1,7 +1,16 @@
 "use strict";
-var WidgetUi = require("../WidgetUi");
 var HelperUtil;
 (function (HelperUtil) {
-    HelperUtil.Ui = WidgetUi;
+    function notNull(obj) {
+        var copy = {};
+        for (var key in obj) {
+            var value = obj[key];
+            if (value != null) {
+                copy[key] = value;
+            }
+        }
+        return copy;
+    }
+    HelperUtil.notNull = notNull;
 })(HelperUtil || (HelperUtil = {}));
 module.exports = HelperUtil;